+++ v3/crates/plan/src/types.rs
@@ -121,6 +121,11 @@ pub enum PermissionError {
     #[error("Error evaluating condition: {0}")]
     ConditionEvaluationError(#[from] authorization_rules::ConditionError),
 
+    #[error("Nested scalar filtering is not supported by data connector {data_connector_name}")]
+    NestedScalarFilteringNotSupported {
+        data_connector_name: Qualified<DataConnectorName>,
+    },
+
     #[error("{0}")]
     Other(String),
 }
@@ -141,7 +146,8 @@ impl TraceableError for PermissionError {
             | Self::FieldNotFoundInBooleanExpressionType { .. }
             | Self::RelationshipNotFoundInBooleanExpressionType { .. }
             | Self::ObjectBooleanExpressionTypeNotFound { .. }
-            | Self::ConditionEvaluationError(_) => ErrorVisibility::Internal,
+            | Self::ConditionEvaluationError(_)
+            | Self::NestedScalarFilteringNotSupported { .. } => ErrorVisibility::Internal,
             Self::Other(_) => ErrorVisibility::User,
         }
     }